feat: permitir horas en medias unidades (ej: 2.5h)
- DB: columnas hours, extra_hours y horas_proyectadas cambiadas de INTEGER a NUMERIC con constraint MOD(x, 0.5) = 0
- Backend: validación cambiada de isInt a isFloat con custom validator de múltiplo de 0.5; parseInt → parseFloat en services
- Frontend: step=0.5 en todos los inputs de horas; parseInt → parseFloat en submit handlers

Migración: 009_fractional_hours.sql

+++ backend/src/services/projectionService.js
@@ -22,7 +22,7 @@ async function crear(body, gestorId, roles) {
     throw crearError('fecha_fin debe ser mayor o igual a fecha_inicio');
   }
 
-  if (parseInt(horas_proyectadas) <= 0) {
+  if (parseFloat(horas_proyectadas) <= 0) {
     throw crearError('horas_proyectadas debe ser mayor a 0');
   }
 
@@ -52,7 +52,7 @@ async function actualizar(id, body, gestorId, roles) {
     throw crearError('fecha_fin debe ser mayor o igual a fecha_inicio');
   }
 
-  if (horas_proyectadas !== undefined && parseInt(horas_proyectadas) <= 0) {
+  if (horas_proyectadas !== undefined && parseFloat(horas_proyectadas) <= 0) {
     throw crearError('horas_proyectadas debe ser mayor a 0');
   }
